% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_data360i.R
\name{search_data360}
\alias{search_data360}
\title{Search data360}
\usage{
search_data360(search_string)
}
\arguments{
\item{search_string}{character: Text string to pass to the search API}
}
\value{
character
}
\description{
Search data360
}
