% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_wdi.R
\name{search_wdi}
\alias{search_wdi}
\title{Search WDI}
\usage{
search_wdi(search_string)
}
\arguments{
\item{search_string}{character: Text string to pass to the search API}
}
\value{
character
}
\description{
Search WDI
}
